Rename files with capitalizations
