+++ packages/Theme/index.tsx
@@ -1,2 +1,2 @@
-export {ThemeProvider, useTheme, withTheme} from './themeProvider';
+export {ThemeProvider, useTheme, withTheme} from './Themeprovider';
 export {DoobooTheme, ThemeType, dark, light} from './colors';
